Refactor. Separate mod el_ast.
