+++ lala_lalrpop/src/lala.lalrpop
@@ -1,4 +1,4 @@
-use crate::ast::*;
+use crate::el_ast::*;
 // use std::str::FromStr;
 
 grammar;
